Gestion des résultats des VPA

+++ module/Consultation/src/Consultation/Model/DemandeVisitePreanesthesiqueTable.php
@@ -8,6 +8,18 @@ class DemandeVisitePreanesthesiqueTable{
 	public function __construct(TableGateway $tableGateway) {
 		$this->tableGateway = $tableGateway;
 	}
+	
+	public function getDemandeVPA($id_cons)
+	{
+		$rowset = $this->tableGateway->select(array(
+				'ID_CONS' => $id_cons,
+		));
+		$row = $rowset->current();
+		if (!$row) {
+			$row = null;
+		}
+		return $row;
+	}
 
 	public function getDemandeVisitePreanesthesique($id){
 		$adapter = $this->tableGateway->getAdapter();
@@ -21,19 +33,42 @@ class DemandeVisitePreanesthesiqueTable{
 		return $result;
 	}
 	
-	public function updateDemandeVisitePreanesthesique($infoDemande){
-		$this->tableGateway->delete(array('ID_CONS'=>$infoDemande['ID_CONS']));
+	public function getResultatVpa($idVpa){
+		$adapter = $this->tableGateway->getAdapter();
+		$sql = new Sql($adapter);
+		$select = $sql->select();
+		$select->from('resultat_vpa');
+		$select->columns(array('*'));
+		$select->where(array('idVpa'=>$idVpa));
 		
-		if($infoDemande['diagnostic'] !='' && $infoDemande['intervention_prevue'] !=''){
-			$donneesVPA = array(
-				'ID_CONS'     => $infoDemande['ID_CONS'],
-				'DIAGNOSTIC'  => $infoDemande['diagnostic'],
-				'OBSERVATION' => $infoDemande['observation'],
-				'INTERVENTION_PREVUE' => $infoDemande['intervention_prevue'],
-				'NUMERO_VPA'  => $infoDemande['numero_vpa'],
-				'TYPE_ANESTHESIE' => $infoDemande['type_anesthesie'],
-			);
-			$this->tableGateway->insert($infoDemande);
+		$stat = $sql->prepareStatementForSqlObject($select);
+		$result = $stat->execute()->current();
+		return $result;
+	}
+	
+	public function updateDemandeVisitePreanesthesique($infoDemande){
+		
+		$demandeVpa = $this->getDemandeVPA($infoDemande['ID_CONS']);
+		
+		$resultatVpa = null;
+		if($demandeVpa){
+			$resultatVpa = $this->getResultatVpa($demandeVpa->idVpa);
+		}
+		
+		if(!$resultatVpa){
+			$this->tableGateway->delete(array('ID_CONS'=>$infoDemande['ID_CONS']));
+			
+			if($infoDemande['diagnostic'] !='' && $infoDemande['intervention_prevue'] !=''){
+				$donneesVPA = array(
+						'ID_CONS'     => $infoDemande['ID_CONS'],
+						'DIAGNOSTIC'  => $infoDemande['diagnostic'],
+						'OBSERVATION' => $infoDemande['observation'],
+						'INTERVENTION_PREVUE' => $infoDemande['intervention_prevue'],
+				);
+				$this->tableGateway->insert($infoDemande);
+			}
 		}
 	}
+	
+	
 }
